add base_project dir. update to cairo 0.5.0
