+++ MANIFEST.in
@@ -1,1 +1,2 @@
 include LICENSE
+recursive-include src/nile/base_project/ *
