+++ src/nile/base_project/Makefile
@@ -1,0 +1,3 @@
+# Build and test
+build :; nile compile
+test  :; pytest tests/
