+++ src/nile/base_project/tests/test_contract.py
@@ -1,0 +1,26 @@
+import os
+import pytest
+
+from starkware.starknet.testing.starknet import Starknet
+
+# The path to the contract source code.
+CONTRACT_FILE = os.path.join("contracts", "contract.cairo")
+
+
+# The testing library uses python's asyncio. So the following
+# decorator and the ``async`` keyword are needed.
+@pytest.mark.asyncio
+async def test_increase_balance():
+    # Create a new Starknet class that simulates the StarkNet
+    # system.
+    starknet = await Starknet.empty()
+
+    # Deploy the contract.
+    contract = await starknet.deploy(CONTRACT_FILE)
+
+    # Invoke increase_balance() twice.
+    await contract.increase_balance(amount=10).invoke()
+    await contract.increase_balance(amount=20).invoke()
+
+    # Check the result of get_balance().
+    assert await contract.get_balance().call() == (30,)
